Half steps

+++ app/src/main/java/com/strummer/practice/library/PracticeLibraryModels.kt
@@ -35,8 +35,8 @@ data class BarChordStep(
     val id: String,
     val songId: String,
     val displayOrder: Int,
-    val startBar: Int = 1,
-    val barCount: Int,
+    val startBar: Double = 1.0,
+    val barCount: Double,
     val chordName: String
 )
